requirements.txt
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
langchain-ollama

llama-index
llama-index-llms-ollama
llama-index-embeddings-huggingface

sentence-transformers
transformers

faiss-cpu
chromadb

rank-bm25

pymupdf
pypdf>=4.0.1,<5.0.0
ollama
ragas
datasets
Get
python-dotenv
pyyaml
tqdm
pandas
numpy
